+++ app_runner/docs/changelog.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## \[Unreleased\]
 
+## Added
+
+- New input type `BfabricAnnotationSpec`.
+    - For now, it only supports one type of annotation, which is "resource_sample".
+
 ## \[0.0.10\] - 2025-01-15
 
 ### Added